feat: Rename gen to commit and refactor commit CLI
